pyproject.toml
[tool.poetry]
name = "nonebot-plugin-mccheck"
version = "0.1.55"
description = "Minecraft服务器状态查询的nonebot插件，支持IPv6/Minecraft server status query plugin for nonebot, IPv6 supported"
readme = "README.md"
requires-python = ">=3.9"
authors = ["molanp <molanpp@outlook.com>"]
license = "GPL-3.0"
homepage = "https://github.com/molanp/nonebot_plugin_mccheck"
repository = "https://github.com/molanp/nonebot_plugin_mccheck"
include = ["nonebot_plugin_mccheck/**/*"]
keywords = ["nonebot", "nonebot2", "minecraft", "i18n", "inquire", "checker", "server"]

[tool.poetry.dependencies]
nonebot-plugin-alconna = ">=0.48.0"
nonebot-plugin-uninfo = ">=0.7.0"
nonebot-plugin-htmlrender = ">=0.3.5"
nonebot2 = ">=2.2.0"
dnspython = ">=2.5.0"
ujson = ">=5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.optional-dependencies]
adapters = [
    "nonebot-adapter-onebot>=2.4.1",
]